fix rotate function

+++ cpp/include/cudf/detail/utilities/hash_functions.cuh
@@ -65,7 +65,7 @@ __device__ inline uint32_t rotate_bits_left(uint32_t x, uint32_t r)
   return __funnelshift_l(x, x, r);
 }
 
-__device__ inline uint64_t rotate_bits_left(uint64_t h, int8_t r)
+__device__ inline uint64_t rotate_bits_left(uint64_t h, uint32_t r)
 {
   return ((h << r) | (h >> (64 - r)));
 }
